Update macos-gnu baselines on Trpd and Jckt subdyn cases
The original baselines for these two cases on macos had been generated using macos 10.14.6 (High Sierra) with gcc-10.2.0, but after upgrading to macos 10.15.7 (Catalina), some of the math libraries in the framework changed.  This resulted in minor differences in the some of the matrix solves in subdyn for these two cases which gives small differences in the loads at two nodes (very small when compared to the magnitude of the loads at the foundation.  These are the only two cases with multiple cross members, and only cases that showed this problem.

This updates so that all cases pass on macos 10.15.7 (Catalina) with gcc-10.2.0 (from homebrew).