Update r-test results for large platform rotation update
Note that the PRP motion input file for the case py_hd_5MW_OC4Semi_WSt_WavesWN is reverted back to the previous version. This is because the Python driver code still assumes the inputs in the PRP motion file are the angular velocity and acceleration, not the time derivatives of the Tait-Bryan angles. In contrast, the FORTRAN HD driver now assumes the PRP input file contains the time derivatives of the Tait-Bryan angles and will convert them to the actual angular velocity and acceleration internally. The outputs of the case py_hd_5MW_OC4Semi_WSt_WavesWN and hd_5MW_OC4Semi_WSt_WavesWN remain the same despite the different input files due to the reason explained above.